refactor: services structure
